Sparrow frontend and backend connection
Created function to allow for the frontend to understand whether or not the backend has the necessary information without having to prompt the user. The frontend calls on the link get-persistent-info which returns two variables in a json format indicating whether the backend has the blackboard link and whether it has the class information.